java/src/test/java/exams/exams3/Exam250Test.java
package exams.exams3;

import org.junit.jupiter.api.Test;

import static org.junit.jupiter.api.Assertions.assertEquals;

class Exam250Test {

    @Test
    void maxProfitAssignment() {
        Exam250 exam = new Exam250();
        assertEquals(100, exam.maxProfitAssignment(new int[]{2, 4, 6, 8, 10}, new int[]{10, 20, 30, 40, 50}, new int[]{4, 5, 6, 7}));
        assertEquals(0, exam.maxProfitAssignment(new int[]{85, 47, 57}, new int[]{24, 66, 99}, new int[]{40, 25, 25}));
        assertEquals(324, exam.maxProfitAssignment(new int[]{68, 35, 52, 47, 86}, new int[]{67, 17, 1, 81, 3}, new int[]{92, 10, 85, 84, 82})); //第一次提交答案错误
    }

}
